Move get arguments to schema converter
