+++ strawberry/schema/schema_converter.py
@@ -1,6 +1,6 @@
 from enum import Enum
 from inspect import isasyncgen, iscoroutine
-from typing import Any, Callable, Dict, Type, cast
+from typing import Any, Callable, Dict, List, Tuple, Type, cast
 
 from graphql import (
     GraphQLArgument,
@@ -24,7 +24,7 @@ from graphql import (
     Undefined,
 )
 
-from strawberry.arguments import UNSET, StrawberryArgument
+from strawberry.arguments import UNSET, StrawberryArgument, convert_arguments
 from strawberry.directive import DirectiveDefinition
 from strawberry.enum import EnumDefinition, EnumValue
 from strawberry.field import StrawberryField
@@ -311,6 +311,37 @@ class GraphQLCoreConverter:
 
         return graphql_object_type
 
+    def _get_arguments(
+        self,
+        field: StrawberryField,
+        source: Any,
+        info: Any,
+        kwargs: Dict[str, Any],
+    ) -> Tuple[List[Any], Dict[str, Any]]:
+        if field.base_resolver is None:
+            return [], {}
+
+        kwargs = convert_arguments(kwargs, field.arguments)
+
+        # the following code allows to omit info and root arguments
+        # by inspecting the original resolver arguments,
+        # if it asks for self, the source will be passed as first argument
+        # if it asks for root, the source it will be passed as kwarg
+        # if it asks for info, the info will be passed as kwarg
+
+        args = []
+
+        if field.base_resolver.has_self_arg:
+            args.append(source)
+
+        if field.base_resolver.has_root_arg:
+            kwargs["root"] = source
+
+        if field.base_resolver.has_info_arg:
+            kwargs["info"] = info
+
+        return args, kwargs
+
     def from_resolver(self, field: StrawberryField) -> Callable:
         def _check_permissions(source: Any, info: Info, kwargs: Dict[str, Any]):
             """
@@ -340,7 +371,11 @@ class GraphQLCoreConverter:
             strawberry_info = _strawberry_info_from_graphql(info)
             _check_permissions(_source, strawberry_info, kwargs)
 
-            result = field.get_result(_source, info=strawberry_info, kwargs=kwargs)
+            args, kwargs = self._get_arguments(
+                field, source=_source, info=strawberry_info, kwargs=kwargs
+            )
+
+            result = field.get_result(_source, args=args, kwargs=kwargs)
 
             if isasyncgen(result):
